requirements.txt
django>=4.2
pillow>=10.0
gunicorn>=21.2
whitenoise>=6.6
